fix: remove circular dep and rename That to This
