+++ projects/ngneat/from-event/src/lib/create-tokens.ts
@@ -1,0 +1,11 @@
+export function createTokens(propertyKey: string) {
+  const event = Symbol(`event ${propertyKey}`);
+  const subject = Symbol(`subject ${propertyKey}`);
+  const subscription = Symbol(`subscription ${propertyKey}`);
+
+  return {
+    event,
+    subject,
+    subscription,
+  } as const;
+}
