+++ projects/ngneat/from-event/src/lib/from-event.ts
@@ -1,15 +1,16 @@
 import { ElementRef } from '@angular/core';
 import { fromEvent, defer } from 'rxjs';
 
-import { createTokens, initIfNeeded, subscribeToEventIfPossible } from './helpers';
-import { That } from './types';
+import { createTokens } from './create-tokens';
+import { initIfNeeded, subscribeToEventIfPossible } from './helpers';
+import { This } from './types';
 
 export function FromEvent<K extends keyof DocumentEventMap>(event: K, eventOptions?: AddEventListenerOptions) {
   return function (target: any, propertyKey: string) {
     const tokens = createTokens(propertyKey);
 
     Object.defineProperty(target, propertyKey, {
-      set(this: That, elementRef?: ElementRef) {
+      set(this: This, elementRef?: ElementRef) {
         initIfNeeded(this, tokens);
 
         if (this[tokens.subscription]) {
@@ -26,7 +27,7 @@ export function FromEvent<K extends keyof DocumentEventMap>(event: K, eventOptio
         this[tokens.event] = fromEvent(elementRef.nativeElement, event, eventOptions);
         subscribeToEventIfPossible(this, tokens);
       },
-      get(this: That) {
+      get(this: This) {
         return defer(() => {
           initIfNeeded(this, tokens);
           subscribeToEventIfPossible(this, tokens);
