+++ projects/ngneat/from-event/src/lib/types.ts
@@ -1,12 +1,12 @@
 import { Observable, Subject, Subscription } from 'rxjs';
 
-import type { createTokens } from './helpers';
+import type { createTokens } from './create-tokens';
 
 export type Tokens = ReturnType<typeof createTokens>;
 
 declare var tokens: Tokens;
 
-export interface That {
+export interface This {
   [tokens.event]: Observable<Event>;
   [tokens.subject]: Subject<Event>;
   [tokens.subscription]: Subscription;
